refactor `cli` module into multiple files
